added it to githubapi instead of github scraping
