+++ packages/graphql_server/src/test/githubApi.test.ts
@@ -1,5 +1,5 @@
-import { getRepositoryTopics } from '../api/githubApi'
-import { getContributors } from '../scraping/githubScraping'
+import { getRepositoryTopics, getContributors } from '../api/githubApi'
+
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 export async function testGetRepositoryTopics(repoFounder: string, repoName: string) {
   //get the real comments from method in hackerNewsScraper. E.g. searchHackerNewsStories('SuperAGI')
